Add H990DS version of panel timings
Despite being named the same from device to device, this file very
clearly DOES differ from device to device.  I'm wondering how badly
broken LGE's development process is due to things like this.  This is
good enough for the display to work on a H990DS with the replacement
bootloader.